pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "change-review-orchestrator"
version = "0.1.0"
description = "Regulated Change Review and Release Readiness Orchestrator for banking SDLC teams"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "Your Team", email = "team@yourorg.com" }]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "structlog>=24.2.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "orjson>=3.10.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
cro = "change_review_orchestrator.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E","W","F","I","B","C4","UP","N","ANN","S","RUF"]
ignore = ["ANN101","ANN102","S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN","S"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
exclude = ["alembic/", "scripts/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=0"
